refactor: describe event trigger in error messages
